Fix provenance test and rename attributeSelector test
